[refactor] PinCount객체 패키지 변경
