+++ src/main/java/bowling/domain/State/State.java
@@ -1,7 +1,6 @@
 package bowling.domain.State;
 
 import bowling.domain.score.Score;
-import bowling.domain.frame.PinCount;
 
 public interface State {
