+++ src/main/java/bowling/domain/frame/FinalFrame.java
@@ -1,5 +1,6 @@
 package bowling.domain.frame;
 
+import bowling.domain.State.PinCount;
 import bowling.domain.score.Score;
 import bowling.domain.State.FinalState;
 import bowling.domain.State.State;
