+++ src/test/java/bowling/domain/StateTest.java
@@ -2,7 +2,7 @@ package bowling.domain;
 
 import bowling.domain.State.Miss;
 import bowling.domain.State.Spare;
-import bowling.domain.frame.PinCount;
+import bowling.domain.State.PinCount;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
